# docs/api/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Marketplace API",
        "version": "0.0.1",
...
    },
    "servers": [
        {
            "url": "http://localhost/api"
        }
    ],
...
    "security": [
        {
            "http": []
        }
    ],
    "paths": {
...
        "/user/avatar": {
            "patch": {
                "operationId": "profile.updateAvatar",
                "description": "Upload an avatar file to update the profile.",
                "summary": "Update avatar",
                "tags": [
                    "User/Profile"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/Requests.UpdateAvatarRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "const": "Avatar updated successfully."
                                        },
                                        "avatar_url": {
                                            "type": "string"
                                        }
                                    },
                                    "required": [
                                        "message",
                                        "avatar_url"
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "const": "No avatar file was provided."
                                        }
                                    },
                                    "required": [
                                        "message"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Requesting protected routes without a valid JWT token, or token is expired/invalid.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error_code": {
                                            "type": "string",
                                            "description": "The global error code.",
                                            "example": "UNAUTHENTICATED"
                                        },
                                        "exception_type": {
                                            "type": "string",
                                            "description": "The exception class name.",
                                            "example": "AuthenticationException"
                                        },
                                        "message": {
                                            "type": "string",
                                            "description": "A human-readable error message.",
                                            "example": "You are not authenticated. Please provide a valid token."
                                        }
                                    },
                                    "required": [
                                        "error_code",
                                        "exception_type",
                                        "message"
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error_code": {
                                            "type": "string",
                                            "description": "The specific error code identifying a validation failure.",
                                            "example": "VALIDATION_ERROR"
                                        },
                                        "message": {
                                            "type": "string",
                                            "description": "Errors overview message.",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "description": "A detailed description of each field that failed validation.",
                                            "additionalProperties": {
                                                "type": "array",
                                                "items": {
                                                    "type": "string",
                                                    "example": "The field is required."
                                                }
                                            }
                                        }
                                    },
                                    "required": [
                                        "error_code",
                                        "message",
                                        "errors"
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Attempting to access a route or perform an action without the required user role or permissions.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error_code": {
                                            "type": "string",
                                            "description": "The global error code.",
                                            "example": "FORBIDDEN"
                                        },
                                        "exception_type": {
                                            "type": "string",
                                            "description": "The exception class name.",
                                            "example": "AuthorizationException"
                                        },
                                        "message": {
                                            "type": "string",
                                            "description": "A human-readable error message.",
                                            "example": "You do not have permission to perform this action."
                                        }
                                    },
                                    "required": [
                                        "error_code",
                                        "exception_type",
                                        "message"
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/user/settings": {
            "patch": {
                "operationId": "profile.updateSettings",
                "description": "Update user-specific settings such as location sharing, nearby, etc.",
                "summary": "Update settings",
                "tags": [
                    "User/Profile"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Requests.UpdateSettingsRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "const": "Settings updated successfully."
                                        },
                                        "settings": {
                                            "type": "array",
                                            "items": []
                                        }
                                    },
                                    "required": [
                                        "message",
                                        "settings"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Requesting protected routes without a valid JWT token, or token is expired/invalid.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error_code": {
                                            "type": "string",
                                            "description": "The global error code.",
                                            "example": "UNAUTHENTICATED"
                                        },
                                        "exception_type": {
                                            "type": "string",
                                            "description": "The exception class name.",
                                            "example": "AuthenticationException"
                                        },
                                        "message": {
                                            "type": "string",
                                            "description": "A human-readable error message.",
                                            "example": "You are not authenticated. Please provide a valid token."
                                        }
                                    },
                                    "required": [
                                        "error_code",
                                        "exception_type",
                                        "message"
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error_code": {
                                            "type": "string",
                                            "description": "The specific error code identifying a validation failure.",
                                            "example": "VALIDATION_ERROR"
                                        },
                                        "message": {
                                            "type": "string",
                                            "description": "Errors overview message.",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "description": "A detailed description of each field that failed validation.",
                                            "additionalProperties": {
                                                "type": "array",
                                                "items": {
                                                    "type": "string",
                                                    "example": "The field is required."
                                                }
                                            }
                                        }
                                    },
                                    "required": [
                                        "error_code",
                                        "message",
                                        "errors"
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Attempting to access a route or perform an action without the required user role or permissions.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error_code": {
                                            "type": "string",
                                            "description": "The global error code.",
                                            "example": "FORBIDDEN"
                                        },
                                        "exception_type": {
                                            "type": "string",
                                            "description": "The exception class name.",
                                            "example": "AuthorizationException"
                                        },
                                        "message": {
                                            "type": "string",
                                            "description": "A human-readable error message.",
                                            "example": "You do not have permission to perform this action."
                                        }
                                    },
                                    "required": [
                                        "error_code",
                                        "exception_type",
                                        "message"
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/username": {
            "get": {
                "operationId": "profile.checkUsername",
                "description": "Check if a username is valid and available (not taken by another user).",
                "summary": "Check username availability",
                "tags": [
                    "User/Profile"
                ],
                "parameters": [
                    {
                        "name": "username",
                        "in": "query",
                        "required": true,
                        "description": "The username to check for availability.",
                        "schema": {
                            "type": "string",
                            "pattern": "^[a-z0-9_]+$",
                            "minLength": 3,
                            "maxLength": 50
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "available": {
                                            "type": "boolean"
                                        }
                                    },
                                    "required": [
                                        "available"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Requesting protected routes without a valid JWT token, or token is expired/invalid.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error_code": {
                                            "type": "string",
                                            "description": "The global error code.",
                                            "example": "UNAUTHENTICATED"
                                        },
                                        "exception_type": {
                                            "type": "string",
                                            "description": "The exception class name.",
                                            "example": "AuthenticationException"
                                        },
                                        "message": {
                                            "type": "string",
                                            "description": "A human-readable error message.",
                                            "example": "You are not authenticated. Please provide a valid token."
                                        }
                                    },
                                    "required": [
                                        "error_code",
                                        "exception_type",
                                        "message"
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error_code": {
                                            "type": "string",
                                            "description": "The specific error code identifying a validation failure.",
                                            "example": "VALIDATION_ERROR"
                                        },
                                        "message": {
                                            "type": "string",
                                            "description": "Errors overview message.",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "description": "A detailed description of each field that failed validation.",
                                            "additionalProperties": {
                                                "type": "array",
                                                "items": {
                                                    "type": "string",
                                                    "example": "The field is required."
                                                }
                                            }
                                        }
                                    },
                                    "required": [
                                        "error_code",
                                        "message",
                                        "errors"
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Attempting to access a route or perform an action without the required user role or permissions.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error_code": {
                                            "type": "string",
                                            "description": "The global error code.",
                                            "example": "FORBIDDEN"
                                        },
                                        "exception_type": {
                                            "type": "string",
                                            "description": "The exception class name.",
                                            "example": "AuthorizationException"
                                        },
                                        "message": {
                                            "type": "string",
                                            "description": "A human-readable error message.",
                                            "example": "You do not have permission to perform this action."
                                        }
                                    },
                                    "required": [
                                        "error_code",
                                        "exception_type",
                                        "message"
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "http": {
                "type": "http",
                "scheme": "bearer"
            }
        },
        "schemas": {
...
            "Requests.UpdateAvatarRequest": {
                "type": "object",
                "properties": {
                    "avatar": {
                        "type": "string",
                        "format": "binary",
                        "contentMediaType": "application/octet-stream",
                        "description": "The avatar image file to upload.",
                        "maxLength": 2048
                    }
                },
                "required": [
                    "avatar"
                ],
                "title": "UpdateAvatarRequest"
            },
...
            "Requests.UpdateSettingsRequest": {
                "type": "object",
                "properties": {
                    "location_sharing": {
                        "type": [
                            "boolean",
                            "null"
                        ]
                    },
                    "nearby": {
                        "type": [
                            "boolean",
                            "null"
                        ]
                    },
                    "sound_recording": {
                        "type": [
                            "boolean",
                            "null"
                        ]
                    },
                    "silent_mode": {
                        "type": [
                            "boolean",
                            "null"
                        ]
                    }
                },
                "title": "UpdateSettingsRequest"
            },
...
        }
    }
}